Move TaskBuilder.ts to tests/TestingTools/
